COMP: HDF5 fails to compile on mingw
The file H5Omtime.c failed to compile under Mingw because the variable
timezone was not defined. The config check set H5_HAVE_TIMEZONE to 1
for all Windows platforms. This patch excludes that set for MINGW.

There is a bit of mystery however. The VTK hdf5 library does compile
properly on mingw and the VTK ConfigureChecks.cmake file is virtually
the same as the ITK version. This indicates there is some other config
difference between VTK and ITK.

Change-Id: I85b2d3c56e9c6a3b3abd6a4b97be63e4166e3b58