ENH: Request only one module at a time
During a modular dashboard we loop over each module to build it as a
subproject dashboard submission.  We enable each module by adding

  -DModule_${subproj}:BOOL=ON

to the configuration line to explicitly request it.  Add also an option
to turn OFF the module enabled on the previous loop iteration.  That way
we configure only one module at a time (plus its dependencies) instead
of accumulating them.