ENH: List module headers and sources in -all targets
We define a <mod>-all target for each module that depends on all the
libraries and executables in the module but does not actually compile
anything itself.  Add to this target the headers and sources for its
module when generating for IDE build tools.  This presents them for easy
editing in the IDEs.

Change-Id: I3fdfe38703f09188b4e3d4e78a5f49f8227ab5a8