COMP: Fix ITK TestExternal build with CMake < 2.8.4
The FindITK module that comes with CMake versions prior to 2.8.4 cannot
find ITKConfig.cmake in the modularized install location under a prefix.
The FindITK module that comes with CMake 2.8.4 and later is just a thin
wrapper around find_package()'s config mode.  Use the config mode
directly to find ITK so that it works with any CMake 2.8 version.