ENH: Removed Examples/Review directory.
In preparation for large scale code-review effort. This directory has been
removed in order to make possible to move the Review directory out of the
toolkit. The content of that directory will then be code reviewed following the
checklist procedure and moved (when appropriate) into their final destination
inside of the toolkit.

The two examples that were in this directory have been copied to
the ITK Wiki Examples at:

http://www.itk.org/Wiki/ITK/Examples/Segmentation/SinglephaseChanAndVeseDenseFieldLevelSetSegmentation
http://www.itk.org/Wiki/ITK/Examples/Segmentation/SinglephaseChanAndVeseSparseFieldLevelSetSegmentation

Change-Id: I6564de8fa55254f7f124e3f827cc21b81a2949d8