ENH: Move <module>-all targets into module directories
Create each module's -all target in the module build directory.  Add all
targets in the module as direct dependencies of the -all target.  This
approach simplifies the creation of the targets and moves the code into
the main modularization infrastructure.  It also allows developers to do

 $ cd Modules/SomeGroup/SomeModule
 $ make

in order to build all the targets in the module (including tests).

Change-Id: Ibf0f9f1a69a8dd87bc67652a442abfa90cccc829