ENH: Remove Examples dependency on Review.
The Examples were depending on ITK_USE_REVIEW for two classes.

          *    itkVTKPolyDataReader
          *    itkMemoryProbesCollectorBase

The first one was moved out of the Review directory recently,
so it didn't need the ITK_USE_REVIEW ifdefs anymore. The second
one was moved into Common before the release of ITK 3.10, so it
didn't need the ifdefs either. In summary the dependency on the
Review directory at this point, was a historic artifact.

Change-Id: Ie6e0ba85fff838a9e430be4d3f6a784337bc4c86