ENH: Enable MaskImageFilter to work with VectorImages.
NumericTraits< T >::Zero is defined as a static member function
when T is a VariableLengthVector, leading to compilation errors
when the MaskImageFilter is templated over VectorImages. This
change moves initialization into the constructor and uses an
overloaded member function to initialize the OutsideValue
variable in a manner appropriate for the output pixel type. The
member function for VariableLengthVectors initializes the
OutsideValue to a zero-length vector.

Also, added a check of the OutsideValue within
BeforeThreadedGenerateData() for VectorImages. If the
OutsideValue is filled with zeros (or zero-length), the
OutsideValue is resized to a zero-filled vector with length equal
to the number of components in the output image. Otherwise, if
the number of components of the OutsideValue does not match the
number of components of the VectorImage, an exception is thrown.

Change-Id: Ib864b74dc92d045a18c7369839123fb546356835