ENH: Update VideoStream to not touch buffered temporal region
Since the implementation of GenerateData in TemporalProcessObject now
handles the updating of the BufferedTemporalRegion for the output, we
should never touch it in TemporalStreamingGenerateData or
ThreadedGenerateData. The assumption for the implementation of these
methods should be that the BufferedTemporalRegion will be updated at the
end of GenerateData to properly reflect that the frames in
RequestedTemporalRegion have been generated. This means that in VideoStream
the empty frames that get initialized in InitializeEmptyFrames must be
taken from the requested region rather than the buffered region since they
have not yet been buffered