ENH: better computation of spacing and origin in HistogramToImageFilter
The origin is in the center of the pixel of index 0 in the destination
image, so it is more logical to define the origin at the same position
than the first measurement vector in the input histogram, instead of the
minimum value in the bin.

The spacing can be computed with a single element in the histogram, simply
by computing the difference between the minimun and maximum of a bin.

Change-Id: I4f1a4673aac36ca3614a1b6c20ba679309f97935