ENH: Teach pre-commit hook to process ExternalData content links
Look for "content links" (e.g. ".md5" files) staged for commit.
Validate the files to ensure they are valid hashes for the corresponding
hash algorithm.  Search for the linked content in three places:

 - ".ExternalData_${algo}_${hash}" file left by ExternalData module
 - ".ExternalData/${algo}/${hash}" file in top-level .ExternalData store
 - "refs/data/${algo}/${hash}" reference to a blob in Git

If the content is found at the first location, move it to the other two.
The second location will keep the object available locally for fetching
by the ExternalData module at build time.  The third location prepares
the object for upload through Git (outside the main project history).