ENH: Restore tests using BrainWeb data
Restore the option ITK_USE_BRAINWEB_DATA that was removed during
modularization.  Restore use of the ExternalData CMake module to obtain
some of the data files at build time.  Fix test command lines to pass
the test executable to the test driver with $<TARGET_FILE:...> syntax.

Some tests depend on the output of other tests.  Previously this
dependency was implicit in the ordering of the tests.  Set the tests'
DEPENDS property to explicitly specify their dependencies so that they
run correctly in parallel testing configurations.

Change-Id: I7f7bdc482128fc50ccbf540246d10f5c2b8c875a