ENH: Move remaining tests out of ITK-IntegratedTest.
Move the rest of tests in ITK-IntegratedTest module
to the modules they should belong to. Some modules' test dependencies
are modified accordingly (comments are added to itk-module.cmake
when tests introduce new test dependencies). A couple of classes
are relocated for better and clearer module dependencies.

The remaining small group of tests are mostly PrintTests, which
need to be removed after creating PrintTests for each module.

The final goal is to completely remove the ITk-IntegratedTest module.

Change-Id: I5bdbaa81b7673342f619aacb3ee7848b21162699