ENH: Add I-Frame safety to VideoFileReader
For OpenCV, we want to make sure that we don't try to read past the
last I-Frame since this will cause reading errors, so we add the
IFrameSafe option to VideoFileReader. If turned on, this option will
use the last I-Frame as the last frame of the video for reporting the
largest possible temporal region.