ENH: Convert "git gerrit-push" alias to a shell script
Replace the inline script in the alias with an invocation of a real
shell script from the work tree.  Write the script to reproduce the
original behavior of the alias.  Teach it to reject pushing changes
without a topic name to prevent this common mistake.

Format the script as a standard bash-implemented Git command.  Implement
basic argument parsing/validation and command-line usage help.  This
will allow additional features to be added later.

Change-Id: Ifeb25e110ce71e9c450607ec607f3e277c363b2f