DOC: Generating graph with dot for module dependencies
First a dot file is generated at cmake time for each module.
Then we generating the doxygen documentation, first dot generates
a png for the graph which is called in the documentation for each
module. Finally the image gets added to the documentation by doxygen.

DOXYGEN_MODULE_ENABLED has been removed, and is now the expected
behavior.

Change-Id: I8609ecb39e93799902a61add4be23acb8555b1b5