ENH: Teach "git gerrit-push" script to push data with topic
Identify ExternalData content links added or modified by commits in the
topic to be pushed.  Convert those that are referenced by refs/data/*
into a commit object and push it to refs/data/commits/$commit in Gerrit.
This approach uploads data needed for a topic through the same
connection used to push the topic.  A separate tool will fetch the data
commit from Gerrit, extract the data blobs from it, and publish them in
a location accessible to clients using the ExternalData CMake module.