ENH: ExternalData: Teach DATA{} syntax to name associated files
Some data files come in pairs or groups such that only one needs to be
named on a command line and the program can find the others next to it.
Support this case with the DATA{} syntax by allowing a comma-separated
list of names or regular expressions to specify associated files.