ENH: Remove itkDebugMacro from itkGet macros
All of the itkGet macros include a reference to itkDebugMacro. Since
itkDebugMacro uses "<<" to print its debug message, itkGet macros can
only be used for types that implement "<<". For example, Get methods
that access std::vector member data cannot use these macros. To
further complicate matters, the itkDebugMacro is only present if
built with debug. This means that a deveoloper building without debug
may falsely conclude that her code will compile when, if debug is on,
it may not.

The DebugMacro was introduced in the Get (and Set) macros primarily
for people running interpreters like tcl or python.

A more comprehensive change would remove itkDebugMacro from all of the
Set macros. But, since many of the Set macros use operator !=, such a
change may not make much of a difference.

Change-Id: Iaa3257df0e8c21a49784d94e0c0249db341e7218