ENH: Changed number of threads type.
Starting in itkProcessObject, m_NumberOfThreads was an int. Since it is a
countable quantity, it should be an unsigned int. Changes to the signatures
of many ThreadedGenerateData and similar functions were also required.
For consistency and documentation a new typedef is introduced here:

                         ThreadIdType

and it is defined as "unsigned int" in the itkIntTypes.h file.

Change-Id: I61acbf0d5c65f007d4d9d3a666f37b15f236cf4d