ENH: make itk.show() able to pass a title to the viewer
By default, itk.show() will try to generate a title to describe the
image.

If the image is not connected to a pipeline, then the class of the
image is copied in the title. For example:

  itkImageUC2

If the image is connected to a pipeline, or if the object passed is a
filter, then the name of the class of the source is copied, with the
output index between []. For example:

  itkImageFileReaderIUC2[0]

If itk.show() is run from the ipython interpreter, then the variable name
used to store the image or the filter is added to the title. For
example:

  reader - itkImageFileReaderIUC2[0]

The title can also be passed manually, by using the Title keyword:

  itk.show( reader, Title=reader.GetFileName() )

Change-Id: I3a911a920971973cf0fd506f77d49f030974035d