BUG:  Mantis bug fix (0012159):  Bspline interpolate non-zero index.
An image with a non-zero index will produce incorrect interpolation
due to the way the ApplyMirrorBoundaryConditions() function was
written.  Changes were made to that function, results were verified,
and I changed the itkBSplineInterpolateImageFuntionTest so that
one of the cases had a non-zero index which failed with the previous
version of the class.

Change-Id: I358fc2ef55ed3f542f4fdeccf91e243edb5a9c55