COMP: Make imported gdcm targets visible throughout ITK
If ITK_USE_SYSTEM_GDCM is on we load GDCMConfig.cmake from the system to
import the gdcm targets.  Since imported targets are visible only in the
directory where they are imported and below, we must import them from
the top-level of ITK so that they will be visible everywhere.

Change-Id: I835ae28f9f8e407463e6f660c11ffef01f8827d9