BUG: Fixed false-negatives on some systems.
In the following code snippet,

int a = ...;
float b = ...;

b = static_cast< float >( a );
b == static_cast< float >( a );

the last line evaluates to false on some systems with some
compilers for some values of a. This change stores the result of
the static_cast in a variable and changes the equality comparison
to use the variable.

Also, improved some helper function names and simplified printing
of type names.

Change-Id: I1c1696bbf89b52b368bbef698e134a748f958abe