ENH: make {Binary,Ternary}FunctorImageFilter work in Physical Space
As previously written, {Binary,Ternary}FunctorImageFilter work only
in index space -- i.e. it assumes both images are the same size,
spacing, and orientation.

This is the normal case, but there are times where it would be
desirable to add two images with different spacing, or even images
whose volumes overlap but are not coincident.

This patch adds the option to use Physical Coordinates to perform the
binary function.  The phsyical location of each voxel in the first
input image is used to extract an interpolated voxel from the second
(and third, for Ternary) image to use as an argument(s) to the
functor.

Change-Id: I683217697f886a266aad785172805b8d7a480161