ENH: MetaDataDictionary is accessible in WrapITK in a more usual way
Adds Get & Set functions to itkMetaDataDictionary, specific language
customization for python and tests for Python and Java.
The customization for python allows access to the data in a more
convenient way in python code. This leverages on python's weak typing
wherein the MetaDataObjectValue can be downcasted in the python code.
This type of access is not possible for other languages like Java.
Therefore, the example for Java uses just the default method to access
the data i.e the Get & Set functions.

Change-Id: I0e50d3e066e9b721ebb8b237054979d262c95c6c