BUG: HDF5 reads/writes bools as int.
Unfortunately, HDF5 reads & writes bools as an int.  Which causes
a problem if you try to pass it a bool variable to read or write.
On write, it will write out 3 bytes of whatever is right after
the bool variable in memory. On read, it corrupts the 3 bytes
right after the bool in memory. So instead I use an int temp var
to read or write.

BUG: If sizeof(long) == sizeof(int), HDF5 can't tell them apart.

While C++ considers long and int different types in the case where
sizeof(long) == sizeof(int), HDF5 doesn't distinguish between
the two.  So if you want long values stored, you need to add
metadata to distinguish between the two types. Same goes for
unsigned and unsigned long.

Change-Id: Id1aeb2d691b525bbd460177ad36a11369b0a6a3a