ENH: Scripts running Valgrind and Coverage locally.
These scripts can be used for running Valgrind dynamic analysis
in a local build, and for running code coverage analysis also in
a local build.

Each script includes instructions on how it should be run.

Fixed configuration of the script for running uncrustify.
It is now configured using CMake, and we have removed from
it the hard coded path from a specific developer.

The path to the KWStyle executable has also been set at
configuration time.

Change-Id: I7f97e5cf3b65bf43924837d694ee6b95d1757bb0