ENH: Added support for ignoring mutli-resolution tiff sub-files
TIFFfiles support a set of 2D images. Some times this is used to
store an image pyramid or multi-resolution images. Peviously, ITK
would assume that the image was 3D and all slices were the same
size. This resulted in a segfault when this type of sub-image was
encountered. This patch ignores these multi-resolution sub-images.

Change-Id: I6dbd9ea33caf9be0df34d5dca0d86fe5a2a09e13