ENH: Improved exception descriptions
All exceptions now have a consistent format and
description. Additional information is provided when possible. For
example, if an fopen for reading or writing fails, the last system
error is included in the description. Most important, the filename is
given in each description.

A number of new exception checks are added. Now, all opj_ calls that
return a value are checked for success.

Change-Id: Iab498df5ce981c652045651e7cf9a1fd3f4f9158