ENH: Code review/Refactoring of the Statistics MembershipFunctions
Refactoring of the Statistics MembershipFunctions classes. Unified the Gaussian
and Mahalanobis membership functions traits and API to use the same trait names
and data types for the mean and covariance.

Major API change is affects the MahalanobisDistanceMembershipFunction. This
function no longer uses vnl_vector and vnl_matrix for the mean and covariance
types. Rather, for the MeanVectorType, it uses a vector type consistent with
the MeasurementVectorType but with an element type that is real. So if the
MeasurementVectorType is a Vector&lt;int,3&gt;, the MeanVectorType is a
Vector&lt;double,3&gt;. The CovarianceMatrixType was changed from a vnl_matrix
to a VariableSizedMatrix to be consistent with the GaussianMembershipFunction.
This can affect how user code calls the MahalanobisDistanceMembershipFunction
API.

Minor API change is the name of the traits used. MeanType was changed to
MeanVectorType and CovarianceType

DOC: enhanced documentation of classes and methods.

Change-Id: I4cec556aba73294a34445453eadeb5fb0c21e0de