ENH: Cleaned up code and added documentation.
Added documentation to explain the difference between ITK and VNL
definitions of forward and backward FFT transforms.  Added
"IsDimensionSizeLegal" method to
itk::VnlFFTComplexConjugateToRealImageFilter to throw an exception if
the size of the image is not a multiple of a combination of 2s, 3s,
and/or 5s in each dimension.

General cleanup of all FFT classes.

Change-Id: Idf13b7212d068340ee72bd7a284d05fbc22fdcca