BUG: KdTree failed to add non-terminal nodes to neighbor list.
This submission addresses the bug described at
http://public.kitware.com/Bug/view.php?id=11411.
I also made the code more in line with proper coding
style.  Also, there were quite a few variables that
were not used and ivar use that made thread safety
a definite issue.  This submission attempts to correct
this so now things should be thread safe.

Change-Id: I4d849c4c6e1603da3d42924fb8822eeb3eefe5ad