ENH: Optionally exclude modules from ITK_BUILD_ALL_MODULES
Add option EXCLUDE_FROM_ALL to a module declaration to specify that the
module should not be enabled by ITK_BUILD_ALL_MODULES automatically.
This is useful for modules that use third-party dependencies not
provided with the main distribution.  We should search for such
dependencies only when the module is explicitly enabled.

Change-Id: Ica219db4b899fe552fa8ef4508702e20f0a80f5b