ENH: Fixing itk::VoronoiDiagram2D
VoronoiDiagram2D is derived from Mesh, however
it could not be used by VTKPolyDataWriter.
We changed the vertex container from std::vector<PointType> to
the m_PointsContainer. All the interfaces were keeped.
The incompatibility is
(1). VertexIterator is MapContainer::Iterator,
which can not be dereferenced by *iter but iter.Value().
(2). CellType is derived from Superclass. Previous
CellType changed to PolygonCellType. Now VoronoiDiagram2D
is able to be used as template argument for VTKPolyDataWriter.

Change-Id: Ie6fa45dab69cb6b3df26896546d3962af172b035