ENH: New classes CompositeTransformWriter and Reader.
NOTE: This patch is resumbission of an earlier patch, merged into
latest itk master branch. These new classes are not final, and are intended
to be used for testing of CompositeTransform in applications.

The CompositeTransform IO will likely take on two forms in the future:
1) Same format as in this patch (master text file that lists separate files
in which CompositeTransform's sub-transforms are written), but made directly
a part of existing TransfromFileReader and Writer classes. This will require
modifying CompositeTransform.
2) An HDF5 version that writes the complete transform including its
sub-transforms into a single HDF5 file.

Simple checks have been added to TransformFileReader and Writer to prevent
use of CompositeTransform in them. These checks can be removed in the
future.

CompositeTransformReader is derived from TransformFileReader.
TxtTransformIO has been modified to also read-in files that
hold composite transform information. TranformFileReader
reads the composite transform file via TxtTransformIO, which
returns a list of TransformBase*. This is used to build an
appropriately-typed CompositeTransform in CompositeTransformReader.

Tests:
CompositeTransformWriterTest - tests just Writer.
CompositeTransformReaderTest - test just Reader.
CompositeTransformWriterAndReaderTest - tests both by writing
output and reading back in, then comparing.

Change-Id: I7943eaa1f20d12c9d0beb449f0dd3dec23cc1a67