ENH: Rename template parameter.
In both LabelMapOverlayImageFilter and LabelMapContourOverlayImageFilter:

The first template parameter was named TInputImage but it was indicating
the type of a LabelMap object. I changed it to TLabelMap.

There were also some typedefs derived from this like InputImageType
that I changed to LabelMapType.

I also added some lines to the documentation to explain the functions to use
to set each input.

Change-Id: Ieb85b9cad0899bacc97aea3eb506cf69ec7b6381