DOC:  Fixed flawed merge behavior during rebase
The change in module names was not properly merged
during the rebase with this documentation fix.
This also uncovered a few ingroup name inconsistencies
that were made consistent.

The additions of new documentation also triggered
a failure in the test to ensure that all documented
classes have a proper "ingroup" identifier.

Change-Id: Ie661580e00bab5ac6541b17b925a808301d78754