ENH: move LabelMap and dependencies out of Review
Some filters, mostly reconstruction related ones, shouldn't be
in that module: they are using LabelMaps internally but are not related
to label map otherwise. This will be adressed in another change.

Test data should be moved in this module. This will also be adressed
in another change.

Thanks to Arnaud Gelas for his contribution to this change!

Change-Id: Ifc40343833458faee1fe926ebf7018b9b729b462