ENH: Reorganizes the WrapITK Libraries to comply with the modular ITK
Dependencies of dependent modules are now automatically resolved.
There is no need anymore to list them explicitly.

Java requires explicit import of packages so it is generated
recurssively automatically.

This single commit is the result of the work of Ashish Sharma, Alex
Gouaillard and me.

Next step: move the directories in the Wrapping/WrapITK/Libraries
directory in the Modules subdirectory and load them from there, using
the ITK modules selected by the user.

Change-Id: I92367f497e58696bb13218ddd9a414a43c939d93