ENH: add BinaryMathematicalMorphology and ImageFusion modules
and move the mathematical morphology classes dedicated to binary image
in BinaryMathematicalMorphology and the classes to colorize the images
for visual inspection in ImageFusion

This allow to move the binary morphology classes implemented using
label maps outside of the LabelMap module.

Remaining mathematical morphology test from the Review directory have
been moved to the ITKMathematicalMorphology module - they are
replacing the quite weak tests used before.

Change-Id: Ie15accbe27f9f54d749ce47357ee771f0acc28b5