ENH: Can modify one label for a given Index in a LabelMap
In the LabelObject, one index can now be removed from a line
which, depending on the configuration, can shorten the
corresponding line, can remove the corresponding line, or
create a new line.

In the LabelMap, there are also several different configurations
when setting the label to a given index (if it was background,
it the new label is background or not, if the label object exists
or not). Some of these cases were missing in the previous
implementation of the SetPixel method.

Thanks to Gaetan Lehmann for his contribution!

It should also close ITK-223

Change-Id: If946188613ec81c8035c131b132ec2b12efd331f