COMP: Get rid of compile warnings by using function overloads.
There were compiler warnings for 'subscript out of bounds' in
the VNL FFT code, because it was using a runtime switch statement
to determine which vnl fft class/method to call.

I moved the code into private helper functions that the compiler
can choose at compile time, so that there will be no code that
appears to go out of bounds in the array size variable.

Change-Id: I33c8fc5c011bc06dd2e18c512c839e5329c96a01