ENH: ExternalData: Optionally debug download step
During object download check for ExternalData_DEBUG_DOWNLOAD in the
environment.  If it is set then leave downloaded files with incorrect
hashes behind in the object database (under their own hash, of course).
This allows users to debug download problems by inspecting the data
returned from a server when it is not the desired object.

Change-Id: I6ffc91804aab9eb7899dd47c06ed216ee1c1bc31