BUG: Fix parallel BuildHeaderTest with existing tree.
A conflict existed during build the header test source and creating the
executable on parallel Windows builds because both could not access the source
file at the same time.

CMake was creating multiple dependencies for each header test executable.  This
is resolved by putting the add_custom_command in the inner loop and having it
only generate a single source file.

Change-Id: Ia3dea60543ab62d63109497356a515efee80dce9