BUG: Add prompt for gerrit-merge with cherry-pick.
When cherry-pick'ing from Gerrit onto a branch that has not changed parents from
the original patch set, gerrit will return a 'no changes made' error.  The
commit sha, however, still changes in the process, so it is not identified in
Gerrit.

A prompt is added during the 'gerrit-merge' check to make it non-blocking.

Change-Id: I02094b2b7dac6096b4fdaa5753e3197fa0fdfdc5