BUG: Remove inaccurate failure of itkImageMomentsTest.
The check on the itkImageMomentsTest would occasionally fail on some
platforms, e.g. MinGW.  This is becaus the computed principal axes were
flipped.  However, this is still a valid result because and eigenvector
multiplied by a constant is still a solution to the eigenvector problem.

A(bx) = lambda * (bx)

where A is the matrix, b is a constant, lambda is an eigenvalue, and x is an
eigenvector.

In this patch we multiply by -1 if necessary in the test.

Change-Id: I7a622e84f78cb8d7c2c8ed5e1dddce9b64c00192