BUG: Remove broken GetJacobian  (ITK-2621)
There was an attempt to preserve the GetJacobian() api in a backwards
compatible way, but the implementation had a few very negative effects
that made migration to ITKv4 very confusing and error prone.

The current solution moves those confusing run-time errors to
compile time errors in ITKv4, and allows custom derived transforms
to be written in a way that is both ITKv3 and ITKv4 compatible.

A complete description of the problem is in the issue tracker.

To make a Transform ITKv4 compliant only, replace GetJacobian()
with GetJacobianWithRespectToParameters().

To make a non-standard transform compliant with both ITKv3 and ITKv4,
move the computation of the JacobianWithRespectToParameters from
GetJacobian() to GetJacobianWithRespectToParameters()
Then add the following delegation code for GetJacobian()

// Compute the Jacobian in one position with respect to parameters
template< class TScalarType, unsigned int NDimensions >
const typename MyTransform< TScalarType, NDimensions >
::JacobianType & MyTransform< TScalarType, NDimensions >
::GetJacobian(const InputPointType & p) const
{
  this->GetJacobianWithRespectToParameters( p, this->m_Jacobian );
  return this->m_Jacobian;
}

Change-Id: I7a55b04e97879946d4994f23d7a7f7451cd3f81a