DOC: Change GetJacobian* to ComputeJacobian*
The transform member functions in ITKv3 used to
compute the Jacobian and store the value on an
internal member variable.  These functions now
compute the Jacobian, but do not modify the class
or any internal variables.

To be more consistent with other ITK member functions
that behave this way, change the names to ComputeJacobian*.

Change-Id: I8c6d86c968f1efb17575d1be5cbad183bf510c62