BUG: Fix UpdateTransformParameters to update m_Parameters.
Transform::UpdateTransformParameters depends on m_Parameters being
up-to-date. However, most transforms store parameters in one or more
other member variables for actual use, and only update m_Parameters when
GetParameters is called. Thus, UpdateTransformParameters now calls
GetParameters before adding its update to m_Parameters. This is
inefficient, but a much easier fix than changing every Transform
to update m_Parameters whenever the parameters change. It was considered
to use the Transform's modified-time to see if a call to GetParameters
was needed, however UpdateTransformParameters calls Modified itself,
thus making such a check useless.

The underlying problem is that Transforms store parameters in one or
more member variables, but then copy these values to the member
variable m_Parameters during a call to GetParameters. On the surface,
it looks like m_Parameters can simply be accessed directly to manipulate
the parameters, when in fact it cannot.

Change-Id: Ie4a6d866c51fc74f70d37ec64605fffc60a1dbb5