ENH: automatic threshold calculators
This is the code of the insight journal contribution "Histogram-based
thresholding - some missing methods".

  http://www.insight-journal.org/browse/publication/811
  http://hdl.handle.net/10380/3279

The Otsu filters and calculators have also been modified to use the
architecture of the other calculators.
The histogram is no more computed by the calculator. This is delegated
to another filter, such as ImageToHistogramFilter. This change lets us
easily implementing the multithreading and the ability to mask the
image to compute the histogram, for example by using
MaskedImageToHistogramFilter or by using the histogram attached to a
LabelObject.

The old OtsuThresholdImageCalculator is available in the
ITKV3Compatibility module.

Patch 15 - Changed itkTypeMacro superclass field from ImageToImageFilter
to HistogramThresholdImageFilter in all subclasses

Removed some redundant include files

added methods for setting number of histogram bins and turning
AutoMinimumMaximum on/off. Still need methods for setting
minimum and maximum explicitly. Not sure of the appropriate
way of doing this with the decorator structures.

Added required number of inputs and outputs.

Patch 16 - Rebased after the removal of the Testing/Data submodule.

Patch 17 - Added my (Richard Beare) complete affiliation. Also changed
the strategy for derived classes that include calculators with different
parameters. This change is intended for discussion. I have removed the
virtual method - ConfigureCalculator - and included a pointer to the
calculator in the derived class. This eliminated dynamic_casts inside the
derived class as well as the local variables in the derived class. However
it leads to the need for explicit Set/Get methods, rather than ones
created using macros. Perhaps there are macros that can help that I am
unaware of? The example is the Intermodes filter.

Note - we also need a new test for intermodes filter using
SetUseInterMode(false)

Change-Id: I697873136fbca86258626284eae5df7b552c6241