STYLE: Removed references to the SGI compiler
There was some legacy code that was needed
to appease the SGI MIPSPro compiler. Since that is
no longer supported in ITKv4, this is house cleaning
to remove unnecessary code.

There were also some comments that indicated SGI
compiler needs that were actually universally needed.
In those cases, the comments were revised to remove
the reference to SGI incompatibility.

Change-Id: I31d5437bf03112f6715553e9897f7258f8beb5bc