ENH:  Added B-spline Displacement Field Transform.
Although ITK already has a B-spline transform, its
use in image registration is less than optimal.
This class encodes one way of speeding up a B-spline
for image registration.

Introduced new integer typedef in the TransformBase class
to store the number of parameters. The use of "unsigned int"
us no longer enough, now that we are supporting deformation
field transforms. This new type NumberOfParametersType is
defined now as IdentifierType and it is the return type of
the GetNumberOfParameters() method.

Change-Id: Iafc1b02706e1f0b446fa0978cca0db90a480432d