BUG: Generalize import/export macros (#ITK-2660)
This adds support for gcc visibility. The new macros ITK_ABI_EXPORT and
ITK_ABI_LOCAL can be used to mark symbol declarations to be always
exported or local.

This is done for class definitions in header files and for ITK
exception classes to mark their corresponding type_info symbol
with default visibility when using gcc >= 4.

ITK Event objects are such classes which must be marked with
default visibility, since they are defined via a macro and
their type_info node would be hidden, if not explicitly set
to default visibility.

Change-Id: I260c9af04b7cd4183f1a5104bfd0a9600d804587