ENH: Configure ExternalData local object stores for ITK
Expose ExternalData option ExternalData_OBJECT_STORES as a cache entry
defaulting to the empty string.  If not set, use a directory in the
build tree as the main store.  Either way, add the .ExternalData
directory at the top of the source tree as a local object store.  Remove
the latter from the URL templates list so that the objects can be used
without copying to another local store.

Change-Id: I0b454872ad4855e28bd03351c41b6cd02f107204