COMP: Intermittent test failure
itkInitializationBiasedParticleSwarmOptimizerTest is failing
intermittenly. This patch enables more verbose (but not too verbose)
output to help track down the problem.

It seems the test was failing a tolerance check on the results of the
optimization. This was due to hitting the max number of iterations
stop condition rather than the parameter tolerance stop
condition. This patch increases the max number of iterations, forcing
a parameter tolerance stop condition.

It is not clear whether this is a deterministic algorithm. Should we
expect consistent answers for each run?

This patch does not address the deterministic issue.

Change-Id: Ie2110c51639f491cb7b8a6287b0eaeb1370c50bd