BUG: Made output and expected input of FFT filters consistent
The ForwardFFTImageFilter's output was different depending on whether
FFTW was enabled. If FFTW was enabled, the output consisted of only
the non-redundant half of the full complex transformed image. If it
was not enabled, the output consisted of the full complex transformed
image. This change makes the output of ForwardFFTImageFilter always
consist of the full complex image, and the InverseFFTImageFilter
always expects its input to be the full complex image.

In addition, the API has changed for the forward and inverse FFT
filters. Because the full complex image is produced, the method
FullMatrix() has been removed from both ForwardFFTImageFilter and
InverseFFTImageFilter.  The methods ActualXDimensionIsOdd(),
SetActualXDimensionIsOdd(), SetActualXDimensionIsOddOff(), and
SetActualXDimensionIsOddOn() have also been removed from
InverseFFTImageFilter.

For access to FFT algorithms optimized for real inputs, two new
filters have been added: RealToHalfHermitianForwardFFTImageFilter and
HalfHermitianToRealInverseFFTImageFilter.
RealToHalfHermitianForwardFFTImageFilter produces the non-redundant
half Hermitian image that used to be produced by ForwardFFTImageFilter
when FFTW was enabled, and HalfHermitianToRealInverseFFTImageFilter
expects that format as input.  Subclasses for Vnl and FFTW are
provided.

In addition, filters for transforming between full and half Hermitian
images are provided.

ISSUE: ITK-2622
Change-Id: I503254e4e99f00df5f259fce1d7b18be40eb2369