BUG: ITK-2698 Introduce NumberOfParametersType
This type replaces the hard coded use of "unsigned int" inside the
code. The new typedef is still defined as an "unsigned int" but now
it can be distinguished from other types.  Also introduced an
ImageDimensionType for the same identification purpose.

The NumberOfParametersType can be a type with capacity smaller than
the number of pixels, it is in general the number of parameters in
the local support. This number tends to be in the range of tens.

Change-Id: I394c21cab8a75559a395b4da25c0dd2b601c1907