ENH: Make the number of headers settable.
This make the number of headers included in the header test settable.  This is
done through the CMake cache variable.

Intermittent bugs still show up because the header order changes randomly.  If
this MAXIMUM_NUMBER_OF_HEADERS is set to 1, it should detect all the the bugs.

The result of a discussion with Bill Lorenson and Arnaud Gelas.

Change-Id: I410632658e9ac89558fc6e488e393e5c78f682bc