BUG: Separated the concepts of "op" and "op="
This is to enable certain operators and filters to operate on Vectors
without producing warnings in the concept checking.

The required operators + and += are quite different. What works for
one may not work for the other. Therefore these concepts have been
separated. Addtionally, static_cast should be avoided in the concept
checking, because the point of these check is to see if ther is a op
for the set of types passed as arguments. If the arguments are casted,
then there is no check if the operator exists for the types passed as
template parameters!

Change-Id: I1f6dccfbd9c2d7cf7662db32221c68c57c321b31