ENH: Adding MD5 check in addition to baseline images
Added command line option to itkTestDriver

 --compare-MD5 TEST md5hash0 [ md5hash1 ... ]
      Compare the TEST image file's md5 hash to the provided hash.
      md5hash0 is required and assumed to be a hash.
      Additional arguments are considerd hashes when the string is 32
      hexi-decimal characters.
      This option can be used several times for multiple comparisons.

For many types of filters we expect the output to be byte for byte the
same for all platforms. Binary Morphology is a good
example. Currently, a baseline image is required. The alternative is
to just check the md5 check some of the data of the images is
consistent. The saves the need to have a baseline image for every
test. When as hash comparison fails only the test image will be
reported to the dashboard.

As the hash is designed to replace the need for a baseline image, when
a test fails the baseline image will not be reported on the
dashboard. Additionally, the hash is just of the pixels, which include
their values, and indirectly the size of the image, it does not
include meta-data such as pixel spacing or  orientation.

This patch is based on work done by SimpleITK development.

Change-Id: I38c5dd6aefdc63d4476928e5ec3034db8f0150ef