ENH: Add reading/writing Composite Transforms
This patch adds the ability to read and write
CompositeTransform into txt-format and HDF5-format
transform files.

This currently supports all transform types supported
by the pre-existing transform reader/writer.  There is
currently no way to incorporate a DisplacementFieldTransform,
but that was already the case.  A future enhancement would
be to incorporate DisplacementField transforms, at least to
the HDF5 file type.

The CompositeTransformReader/Writer classes in Review
were removed.

Change-Id: I9e470f471bd681f6342759959ff7bb844cba8578