ENH: Add tolerance to check for identity scales.
ObjectToObjectOptimizerBase - in check for identity scales in
StartOptimization, allow values within tolerance of 1 to be considered
1. This provides for numerical inaccuracies during automatic estimation
of scales.

Current the scale values are left unchanged when they are within tolerance,
and a flag is just set to indicate identity scales. This seems reasonable
but could be changed to replace the values with exactly 1.0
if there's an argument for it.

Change-Id: I4e518bc6e2343bbf743ec738c26803fe52db6159