COMP: fix broken region comparison when the dimension is modified
The operator== for ImageRegion works only for the regions of same sizes.
As a consequence, attempting to build InPlaceImageFilter with input
and output of different dimension was trigerring a build error.
The regions are now compared without using the operator== problem.

COMP: Add missing types required by the new class hierarchy of
ExtractImageFilter

Change-Id: Icdf79f86fab37250fbfa58bb36bdef42f0107233