COMP: Explicitly set public/static/const/virtual
This patch set addresses concerns identified by the intel compiler
regarding const correctness, implied public interfaces,
functions that should be static to limit their scope, and
virtual base class destructors.

The base class destructors are not virtual on purpose for performance
reasons.
remark #444: destructor for base class "vnl_vector<>" is not virtual
remark #444: destructor for base class "vnl_matrix<>" is not virtual
remark #444: destructor for base class "Array<>" is not virtual

This is not a common case, and it is desired that this warning be
emmitted in most cases.

COMP: Add virtual destructors vnl_svn

COMP:  Avoid remark on well tested template code in itkByteSwapper
remark #280: selector expression is constant

Change-Id: I13595f1ba016d5f30852f82c624a03c5a4833b2a