BUG: Set LevelSetsv4Visualization libs for VTK 5.6.
In VTK 5.6 is used, it does not set the VTK_LIBRARIES variable, and the
libraries for linking have to be specified explicitly.
Here, we specify these libraries.

We also populate the ITKLevelSetsv4Visualization_LIBRARIES variable for users of
the module.

Finally, we prevent the ITKVtkGlue module from clobbering the VTK_LIBRARIES
variable.

Change-Id: I7c6081ac438c60d4ec43350add7b64e08bb23b21