BUG: RegisterDefaultTransforms not called until too late
I noticed that in the TransformBase tests

itkTransformFactoryBaseTest.cxx calls
itk::TransformFactoryBase::RegisterDefaultTransforms();

The other tests (for Reader & Writer) do not.

Some of the tests in BRAINSResample were failing, claiming that the
ObjectFactory couldn't create an instance of AffineTransformDouble_3_3.

If I add itk::TransformFactoryBase::RegisterDefaultTransforms(); to the
main function, the BRAINSResample tests succeed.

Without it, RegisterDefaultTransform gets called from
itk::TransformFactoryBase::GetFactory(), but only after trying to call
ObjectFactoryBase::CreateInstance -- down in the exception code!

I'm trying out a patch, but it depends on the call toe
TransformFactoryBase::GetFactory() having the side effect of registering
all the transform types. This is disturbing to me because it depends on
side effects!

So it seems like there's a few problems with the Transform Factory: The
initialization isn't thread safe, and it isn't getting done before it's
needed.

Isn't there a better way to make sure that RegisterDefaultTransforms gets
called?

Change-Id: I6ca5c28265175086df53da52a1a7664ff6eb0d13