BUG: Opt metrics work with both BSpline*Transform
The optimized paths for BSpline*Transform types
need to work with both BSplineTransform and BSplineDeformationTransform
types.

This bug was exposed by the test suite of BRAINSFit within the
Slicer application.  Registrations were taking 2hours instead of 10
minutes, and the results were not good.

This appears to make the registrations run quickly again (at least as
quick as when build against ITKv3), and the results are the same as
when building against ITKv3.

Change-Id: I028e53dac012c72999974323d687d831c76cda47