BUG: Added output image template parameter to boundary conditions
The boundary conditions were templated only over the type of image
for which they provided values. This change adds a template
parameter specifying an output image type to which the values
retrieved from the boundary condition will be stored. The pixel
accessor methods in the boundary conditions now return the pixel type
of this output image rather than the pixel type of the input image.
This solves some compilation problems when the input and output image
types are different in PadImageFilter and its subclasses.

Tests for the pad image filters were modified to improve style
as well as exercise the filters when the input and output image
types are different.

The ConstantPadImageFilter is now marked as modified if the constant
value is changed.

Change-Id: I0192e62cc1677fa5d7158101f440d02d5d7e8137