COMP: Removed unreachable code.
This also implements the "Prefer
initialization to assignment" rule.

warning #177: variable "" was declared but never referenced
warning #593: variable "" was set but never used
remark #185: dynamic initialization in unreachable code
remark #111: statement is unreachable

The intended behavior of the preprocessing
code was not properly implemented, and the
code was never executed.

remark #193: zero used for undefined preprocessing identifier "TDimension"

Change-Id: I8a4840e5ad248202b22b0d57fc44b5a3a9594dbb