ENH: Add release helper script SourceTarball.bash
Load the source tree for a given commit.  Find ExternalData content
links in the tree, download the referenced data objects, and load them
into path ".ExternalData/<algo>/<hash>".  Store the result as a Git tree
object and use "git archive" to generate either a tarball or zip file.
Use LF newlines for the tarball and CRLF newlines for the zip file.

Change-Id: I480e0ae92075ab85f7d273649d3a9252b3558e53