ENH: Check in pre-commit that developer setup is up-to-date (ITK-140)
Record a SetupForDevelopment.sh script version in the local git config.
Check the value during pre-commit to be sure it is not smaller than the
current version of the script.  It is difficult to test that all the
effects of development setup are in place, but this sanity check will
avoid accidentally out-of-date trees.

Change-Id: I8e6d8e8ec89cd792d4d2144621481caf6d03841c