COMP: Do not include ITKExternalData explicitly in FEM/test (again)
Commit 23f10a98 (FEM refactoring as contracted for ITKv4, 2011-06-29)
accidentally added an explicit inclusion of ITKExternalData in the FEM
module's test directory.  Direct inclusion does not work for ITK
TestExternal builds.  Since the CMake module is included from the
top-level source directory it is not necessary to include it here.

Restore fix from commit b4b964be, the original version of this commit.
It was un-fixed by commit 8c742085 (ImageToRectilinearFEMObjectFilter
Bug Fix, 2011-09-28).

Change-Id: I7bbec41c30d616574be5471fc8d11772652b9136