ENH: Move period counting to the LevelSet Command.
A member is added to LevelSetEvolution base to get the number of iterations.

The level set command now contains a period setting to set the period of
Update() calls.

Since it relies on the presence of a GetNumberOfIterations method, the command
is renamed to LevelSetIterationUpdateCommand.  Note, that the Optimizer classes
also have a GetNumberOfIterations, though.

Change-Id: I658ba36ef405f39e7b32462e90089a6a78d956aa