ENH: use ITK modules dependencies in wrapping
This fixes the broken dependencies in ITKHighDimensionalMetrics.

COMP: fix a few missing base classes

ENH: move convolution filters to the ITKConvolution module

BUG: fix HuangThresholdCalculator name in doxygen comment
It was preventing wrapping to build properly with documentation.

BUG: fix doxygen comment in InitializationBiasedParticleSwarmOptimizer
Not sure why but the \f[...\f] was preventing the documentation to be
built for this class during the wrapping build.

Change-Id: Id575a28b839e05aa76b2ff77a84e9256062a1bc1