ENH: Made the KernelFunction base class templated
It is currently hard-coded to type double, but in some cases it makes
sense to use single precision for these kernel operations.
This was identified in the itkMattesMutualInformationImageToImageMetric
where the kernel is repeated computed in double precision, but then
typecast the input and the outputs to be double from the more
naturual single precision.

Expose a common RealType typename from all classes uniformly.
Use ConceptChecking to enforce floating point data type.

Made itkKernelFunctionBase.h as the ITKv4 templated version,
and added itkKernelFunction.h to the V3Compatibility directory
to provide the old interface.

ISSUE: ITK-2848
Change-Id: I8f39b7aa3e7fec9641eabd0849d599ea80e6401f