PERF: Updated lbfgsb to fix bug
http://users.eecs.northwestern.edu/~nocedal/lbfgsb.html

Condition for Use: This software is freely available, but we expect that
all publications describing  work using this software , or all
commercial products using it, quote at least one of the references given
below. This software is released under the BSD License. This note has
been added to the NOTICE file at the top of the ITK source tree.

This remark describes an improvement and a correction to Algorithm 778.
It is shown that the performance of the algorithm can be improved
signiﬁcantly by making a relatively simple modiﬁcation to the subspace
minimization phase. The correction concerns an error caused by the use
of routine dpmeps to estimate machine precision.

Change-Id: Iee35b8c6ee383b27a9870f73edbb604508c6ef2f