BUG: GaussianSmoothingOnUpdateDisplacementFieldTransform mem leak.
Valgrind reports that memcpy calls have the same source and destination.

We switch to the strategy used in
GaussianSmoothingOnUpdateTimeVaryingVelocityFieldTransform for buffer swapping,
which is simpler.

Also fix a bug where the zeroing on the update bounds assumed a zero start index
in the image.  Use the code in
GaussianSmoothingOnUpdateTimeVaryingVelocityFieldTransform, but fix a bug in
both where a minus one was missing from the end index.

Use ImageAlgorithm::Copy for the copying now in both
GaussianSmoothingOnUpdateDisplacementFieldTransform and
GaussianSmoothingOnUpdateTimeVaryingVelocityFieldTransform.

This fixes the valgrind memleak.  Also, the test observer had to be commented
out because more memory leaks were being reported in the Nifti writer (?).

Change-Id: Ib353341ce5ab9cc722361791eb419f8cd638ade8