ENH: Add ThreadedIteratorRangePartitioner.
Partitions a domain specified by two iterators, e.g. ITK iterators or STL
iterators, into sub-domains for multi-threading.

First it walks through the entire domain to see how many elements are present,
then it finds the iterators for a sub-domain that split it into relatively equal
sized partitions.

This process is slower than an ThreadedIndexedContainerPartitioner, but it is
general enough to be applied to containers like std::map or std::list.

The ThreadedIndexedContainerPartitionerTest is adjusted for this class.

Change-Id: I34d14e54a0931b58f229616b3cc0c63423b635cf