COMP: Be compliant with ANSI standard
There a a GNUC compiler optimization that
is being used by the clang compiler and
issuing many warnings.

Only integer types can be initialized for static class member
variable at declaration time.  The use of floating point
initialization is GNU extension (and is now part of the c++11
ansi standard).

Change-Id: Ic06efaff0654f9d320a51a83e0daee431cbf3e38