BUG: Fix mismatch in internal image types for IsolatedWatershed
Promoted the intermediate pixel type used for the gradient magnitude
to Real. With the output just being 3 values for the 3 labels, it is
quite sensable to want a 8-bit integer as output.

The regression image changed so a new baseline was added.

Also uncovered a bug in the binary search algortihm. Now if the final
guess still results in connected basins, the previous lower guess is
used.

Change-Id: I5614682bbab34003652e2f4432bb44d3b882cb13