COMP: Unterminated string causes valgrind defect.
NOTE: these changes were committed in Dec 2010 to the previous version
of lbfgsb.

The netlib s_copy procedure copies n characters from one string into
another string. Just like strncpy, the null termination character is
not copied unless it is within the first n characters of the source
string. In lbfgsb.c, s_copy is used to updated a char[60] variable
called task. This variable, when used in a printf statement, will
cause a defect:

"Conditional jump or move depends on uninitialised value(s)"

This patch adds 1 to each s_copy(task,...) call so that the null
termination from the literal string will be included in the copy.

Change-Id: Ib1753e61b5ca9d2a066d09cbb1d124321c017707