ENH: Adding Image::Rebind to support smoothing vector images
This patch enables the SmoothingRecursiveGaussianImageFilter to work
with Image, Image's of Vector-like pixels, and VectorImages.

There is a significant problem of being able to change the pixel type,
or the image dimension across all image-types that ITK has. This patch
introduces a member structure into image classes. The Rebind
meta-programming object enables portably the changing a individual
template parameter in an ITK image-type.

Change-Id: I61542667a48b6c0f803ac84ad104728bf0bd2027