BUG: Restoring adaptor support to SmoothingRecursiveGaussian
DOC: fixing rebind documentation in Image class

This patch allows ImageAdaptors to be used with
SmoothingRecursiveGaussianImageFilter. An explicit test with the
filter for image adaptor has been added. This is accomplished by
adding a Rebind class to ImageAdaptors.

All the ImageAdaptors were review for compatibility. The only adaptor
which does not have a scalar output is RGBToVector and VectorToRGB
adaptors. As the Rebind works with the Image class it will also work
here.

Change-Id: I12e7ca3a78d4b6d12f1d8bf7be29ccd283873f39