BUG: VotingBinary could not set output pixels before
If the input pixel was neither a foreground value nor a background
value. The output pixel would be unset. Now the value from the input
is just carried through to the output. This is consistent with the
documentation of the Birth and Death rate IVARS.

Additionally, cleaned up the logic to make the cases less convoluted.

Rewrote test to intantiate filter with more type, and actually have
the output checked instead of just printed to stdout. Also use real
input instead of a random image.

Change-Id: I6d54e2cb962161e5beb26b925268dd3135b1db70