BUG: fix GPUCommon tests
This commit is the culmination of work that Won-Ki started to fix and
improve the GPUCommon code.  This work was done for a while in
graphor's github account and then in krismz's github account.  Part of
this effort is to organize the GPU code into modules corresponding to
the module structure already existing in ITK ie:
   Modules
      Core
        Common
        GPUCommon
        FiniteDifference
	GPUFiniteDifference
	...
      Filtering
        ImageFilterBase
	GPUImageFilterBase
	Thresholding
	GPUThresholding
	...
      ...
There will be a series of commits where each GPU module is introduced.

Also, there was some work done to have the build process copy the
source code from the .cl files into strings accessible at run time so
that the .cl files do not need to be packaged and distributed.

For some reason combining a file rename (from Modules/GPU/Common to
Modules/Core/GPUCommon) with an edit of that file in this same commit
causes git to think that one file is deleted and a new file created.
Thus the diffs from this process are unhelpful to show the actual
changes.

Change-Id: I9eeb4ccaac306500f07cc963a883167060e85208