COMP: global Clone() architecture compatible with java
A Clone() method was recently introduced for the Transform classes,
but this implementation is unfortunately incompatible with java,
because the implementation in each transform returns a different
type than in its superclasses.

Several other implementations of Clone() methods are used in the toolkit, so
there is definitely a need for that.

This change provide a global implementation of the Clone() method for
all the subclasses of LightObject. As in the current implementation
dedicated to transforms, it returns the actual type of the cloned
object. Because the method is available in LightObject and it
subclasses, the current java infrastructure can be used to reimplement the
Clone() method in all the subclasses of LightObject in a java
compatible way.

Change-Id: I92a558fe7ad8e13e219802b92ca68f96d101972e