PERF: Switched order of if-statement and for-loop in DetermineRegionOfSupport
And also rewrote the function a little bit for readability.

Switching the order of the if-statement and the for-loop showed
performance enhancements on a win7, VS2008 setup. On Linux Ubuntu
with gcc 4.4.3 no performance change was noticed: apparently gcc
is quite smart.

https://itk.icts.uiowa.edu/jira/browse/ITK-2889

Tested within the elastix software to give identical results.
All ITK tests still green:
http://www.cdash.org/CDash/viewTest.php?buildid=1988027

Change-Id: If638e0ee5036833b9495248f5a642fa1f80792d7