BUG: Errors from dashboard
dashboard testing has produced warnings from valgrind and failed tests
on two filters for ubuntu/debian machines with g++ 4.1 and 4.4.

The valgrind warnings derive from in correct use of a macro within a base
class constructor (HistogramThresholdImageFilter).

My best guess at the other errors, not having an equivalent machine to
test on, is a minor difference in floating point precision. I'm not an
export on portable floating point code, so I'm open to advice here. The
failures appear to result from differences in estimated threshold of 1
histogram bin. The rule for setting the threshold has a > test of new vs
old entropy in both of these filters. On my test machine (mac, gcc 4.2.1)
the two values appear exactly equal. I presume this isn't the case on the
test machine. I've included a tolerance to see whether it leads to passing
the tests. I'd hope there is a better solution.

The changes that strengthen the floating point checks in the
KittlerIllingworthCalculator were raising exceptions that should
be warnings. The test image has a spike in the histogram at 255 -
this means that one of the denominators becomes zero.
The correct response is to issue a warning and set the threshold
to the previous estimate.

Change-Id: Ibd02b86ccfabab09c1edd4c5ccd20411d8609fe5