BUG: corrected libtiff to be confgured for big-endian system
Prior, a "#undef WORDS_BIGENDIAN" was left unchange from the autoconf
system. This in combination with the confusing structing has not been
simplified. This preprocessor variable is now defined for bigendian
system and undefine otherwise.

There was one instance of the tiff library where the value of the
preprocess definition was check not it's definition. This was changed
so that cmakedefine could be used.

Change-Id: Id4fe91a498c715ca174974dc4e31cd4d87ea3ce4