ENH: adding additional dart metric for image error statistics
Adding the name of the file tested for the ImageError metric.

Additionally compute the minimum and maximum of the pixels that exceed
the tolerances specified.

The mean difference was incorrectly divided by the total number of
pixel in the region, while only sumed over the pixels which exceeded
the tolerance. So it was neither the mean of the whole image nor the
mean of the pixel that exceed the tolerance. Therefore, it was made
consistent with the other statistics that describe the pixel that
exceed the tolerance.

Change-Id: Ia1f2e58e8a0c26b5a86a3840495f6e1c681c2fd5