ENH: Don't use pretty functions strings, as they cause bloat
The pretty functions strings can add a large amount of size to a
library because they are unique for each template parameterization in
a function, or member function. This is commonly used with the
exception macro, and there are a very large number of uses of this
macro. For example SimpleITK had over 38 thousand pretty function
strings adding over 10 megabytes.

This patch disables the usage of prettty functions string, when NDEBUG
is defined. This pretty information is redundent with the line number
and file name, which scales significantly better templated functions.

Change-Id: I0fd310fb5f29bfccaf30373884ff562c07a4767f