BUG: Fixed template arguments in deconvolution filters
The deconvolution filters were not specifying the correct template
types when defining their superclasses, making it impossible to
instantiate some of the classes when the input, kernel, and output
images were different types.

Fixed compilation problems when the internal working precision
was set to float instead of double.

The tests were modified to exercise these fixes.

Change-Id: I53f566032c48d14e34244bc3c434bc2fd4219a47