BUG: Fix exporting for Nrrd library
The Nrrd library was always marked for dllexport due to a defines. In
actuallity, on windows this library is build as static. Now
"TEEM_STATIC" is correct defined so that nothing is marked as a dll
interface.

Additionally, the NrrdIO.h file was removed from the itkNrddImageIO.h
to improve encapsulation, and and reduce unneeded headers being used.

Change-Id: I2e23050928cdb9fe6779c29c50a36ab7f4093c9f