COMP: ITK-2901 arg_min is not a member of vnl_sse<double>.
The vnl SSE2 implementations of arg_min and arg_max for float and double had
never been available.  This results in a compile time failure when the option
VNL_CONFIG_ENABLE_SSE2 is ON.

New implementations for these functions are created.

Also fix a bug in the calculation of max() where it did not give the correct
value a negative maximum entry.  Additional tests are added to verify correct
behavior.

Change-Id: I6788227816d7e08f50a925e71fc515f5cc74f55d