ENH: For VectorImage support, iterators must not return by reference
This patch changes return by referecnce to return by value. The
Image::GetPixel method returns by reference while the
VectorImage::GetPixel method returns by value. A reference to the
temporary returned by the VectorImage methods can not be used.

These ConditionalIterators now follow the same convention as the
ImageRegionIterators. If return by reference is required later, then
the &Value() methods should be used to conform to the other iterators.

Pratically this patch is needed to enable VectorConfidenceConnected to
support VectorImages so that SimpleITK can use this filter.

Change-Id: Ib7b7d7e7810ee93254ecbacc0fbc410ac374d108