DOC: correct documentation of Directions
Main component of this patch is to correct the documentation of the
Directions.  It is the columns of the Directions that provide the
direction traversed by each dimension.

Also performed some general cleanup of the documentation.

Also moved the method SetRegions() out of the subclasses of ImageBase
and put it directly in ImageBase. All the implementations were identical
with the exception of some variation of const-ness of the parameters.

Change-Id: Id340b4bb66d20c79b4dd825853b4f74649cdd12e