ENH: XML DOM for ITK preview 5.
ITK provided functions to read objects from or write objects to XML files.
However, the included XML parser uses the SAX (Simple API for XML)
technique, which makes implementing user XML-based readers/writers quite
difficult and error-prone. This is because the information are pushed to users
during parsing, and the users have to keep the incomplete intermediate states
and instantly handle every piece of incoming information.

This patch adds DOM (Document Object Model) based XML reading/writing
functions to ITK, which greatly simplifies the processes of XML-based
object reading/writing. In this I/O model, reading/writing is with respect
to an intermediate DOM object that represents an XML file. The advantages are
that pulling/pushing information from/to an ITK object is easy, and
reading/writing a DOM object from/to an XML file is performed implicitly.
Information are cached in the intermediate DOM object, so the user doesn't
have to track the states during the reading/writing process.

This patch includes DOM-based add-ons to ITK library ITKIOXML, testings
for the implementation, testing data (XML files), and examples that
demonstrate the use of this new feature.

Change-Id: Ibfacbac653e32f20ff1784d4a4349bed3cd92a56
Signed-off-by: Ren-Hui Gong <rhgong@gmail.com>