COMP: Enable an Iterator with VectorImageToImageAdaptor
There was a compile error when trying to instantiate an ImageIterator
with ImageType of VectorImageToImageAdaptor.

A solution using the current infrastructure, by having the
ImageAdaptor use the DefaultVectorPixelAccessorFunction, and adding
the input with offset methods to the VectorImage adaptor.

A complication a rose with when the PixelAccessor should be updated
with the number of components in the image. The SetVectorLength
method of the accessor is not called when SetImage is call and when
UpdateOutputInformation is call. Simular to other image attributes, if
the user directly updates the interal adapted image, they may not be
propagated to the ImageAdaptor.

Change-Id: If9db42b7422dc78ce256f79361d81776692ede60