ENH: adding VectorImage support to Gradient image filter
A fourth template parameter was added so that the output image type
can be specified. It defaults to the Image of CovarientVectors which
was previous used.

Additionally an overloaded method was needed to correctly transform
the VariableLengthVector, however the code remains the same for the
Covariant type.

A new test was added which takes an input image and computes the
gradient with covariant and vector image output. The output is
compared to a baseline image.

STYLE: style changes were needed to the gradient test

Change-Id: Idb89f0856a22e143117a6221e1e7a375e9fb0ade