PERF: refactor SliceBySliceImageFilter for performance
Utilize the default implementation for GenerateInputRequestedRegion,
which just copies the enlarged output requested region to the input
images. Modified the EnlargeOutputRequestedRegion methods to enlarge
all dimension to the largest except m_Dimensions, so that the filter
can stream in the slice dimension.

The vectors of image internalInputs was being allocated and the image
initialized each loop. This was relocated to outside the loop, and
leaving the Image::Allocate methods inside the loop so that it will
still be reallocated if needed for in-place filter.

The coping for the image slices is now down with the
ImageAlgorithm::Copy method. The should be more efficient then the
prior implementation and more readable. Additionlly, it specifies the
input and output regions, to clearly describe the regions being
copied. As side effect the costly GetOutput method is no longer user
on a per pixel basis.

Change-Id: I7e0f6b244b95529435a9d8c30d5f445997c9da9d