COMP: remove erouneous SWIG version check
This version check code should never be run. If the user provides a
SWIG version, CMake version >2.8 is required by ITK and this code
should not be run. On the other hand if swig is compiled as an
external probject, then we know the version and don't need to check.

Change-Id: Ia01db827682b392dfb4b31f370849992eb250a41