COMP: Avoid C++0x constexpr error
In class definition of not integral type is technically not allowed by
C++03, however some compilers support this with an extension.  Other
compilers provide support, but issue compiler warnings.  VNL uses
a try compile to detect this feature and uses it to take advantage of
optimized constant expressions. This patch changes the constants defined
in vnl_math from class scope to a name space scope.  The usage syntax
will remain identical, but the simplified constant expressions are
uniformly understood (without compiler extensions) accross a much
wider range of compiler dialects.

Change-Id: I360269c2ccff074098fe0242756ffa43e77ea414