BUG: Fixed show2d WrapITK command for Macs.
To properly launch ImageJ on Macs it needs to be invoked with the
open command, and we need to use a macro that loads the image,
sets the view and window title.  For label images, the macro also
tells ImageJ to run the "3-3-2 RGB" command to map gray values
to color.

Also, because of this invocation method, open is given the "-n"
flag specifying a new instance of ImageJ every time.  A previously
running instance of ImageJ doesn't receive the command-line macro
properly.

Change-Id: I4fbd875a76759437c94a2f6dad2fb7c0c8b41880