COMP: Fix wrapping for dimension = 1.
A number of wrapped filters could not be instantiated when a wrapping dimension
of 1 was specified.

Many of the filters, e.g. BinaryContourImageFilter, in SetupLineOffsets(),
instantiate a helper image time whose dimension is one less than the Image
dimension over which the filter is specified.

We restrict the instantiation of filters that could not be instantiated at this
time in the wrapping code to "2+".

A minor correction is also made in the comments of the wrapping code of where to
place the dimension restriction in the arguments of the CMake wrapping macros.

Change-Id: Iff3136f19b5aeb003901c4a540f4a91ca655ecdc