BUG: Using a pointer to data that goes out of scope
itkFEMElement3DTest was failing on gcc 4.7 platforms. The test was
assigning a pointer to data that went out of scope before the pointer
was used. This data was the expected solutions for a number of test
data sets.

The patch moves the expected solutions data declarations into the main
scope.

Change-Id: I5d2d5e794a68a08f80493a7d1c96144bc0a7bd1d